+++ .serverless/cloudformation-template-update-stack.json
@@ -104,10 +104,10 @@
           "S3Bucket": {
             "Ref": "ServerlessDeploymentBucket"
           },
-          "S3Key": "serverless/jobget/dev/1507398194253-2017-10-07T17:43:14.253Z/jobget.zip"
+          "S3Key": "serverless/jobget/dev/1507410257232-2017-10-07T21:04:17.232Z/jobget.zip"
         },
         "FunctionName": "jobget-dev-createCompany",
-        "Handler": "jobget-service/app.createCompany",
+        "Handler": "api/app.createCompany",
         "MemorySize": 1024,
         "Role": {
           "Fn::GetAtt": [
@@ -129,14 +129,14 @@
         "IamRoleLambdaExecution"
       ]
     },
-    "CreateCompanyLambdaVersionDCkSykzItO13HGZDf77sYLdjwclQ98qIcwS4Ytmy8A": {
+    "CreateCompanyLambdaVersionbOhoKJdCbePXTi0Rpo1weBy21MBhxWjw2NrNVy0": {
       "Type": "AWS::Lambda::Version",
       "DeletionPolicy": "Retain",
       "Properties": {
         "FunctionName": {
           "Ref": "CreateCompanyLambdaFunction"
         },
-        "CodeSha256": "DCkSykzItO13HGZDf77s+YLdjwclQ98qIcwS4Ytmy8A="
+        "CodeSha256": "bOhoKJdCbePXTi0Rpo1weBy21M/B+hxWj/w2NrNVy+0="
       }
     },
     "CreateSubscriberLambdaFunction": {
@@ -146,10 +146,10 @@
           "S3Bucket": {
             "Ref": "ServerlessDeploymentBucket"
           },
-          "S3Key": "serverless/jobget/dev/1507398194253-2017-10-07T17:43:14.253Z/jobget.zip"
+          "S3Key": "serverless/jobget/dev/1507410257232-2017-10-07T21:04:17.232Z/jobget.zip"
         },
         "FunctionName": "jobget-dev-createSubscriber",
-        "Handler": "jobget-service/app.createSubscriber",
+        "Handler": "api/app.createSubscriber",
         "MemorySize": 1024,
         "Role": {
           "Fn::GetAtt": [
@@ -171,14 +171,14 @@
         "IamRoleLambdaExecution"
       ]
     },
-    "CreateSubscriberLambdaVersionDCkSykzItO13HGZDf77sYLdjwclQ98qIcwS4Ytmy8A": {
+    "CreateSubscriberLambdaVersionbOhoKJdCbePXTi0Rpo1weBy21MBhxWjw2NrNVy0": {
       "Type": "AWS::Lambda::Version",
       "DeletionPolicy": "Retain",
       "Properties": {
         "FunctionName": {
           "Ref": "CreateSubscriberLambdaFunction"
         },
-        "CodeSha256": "DCkSykzItO13HGZDf77s+YLdjwclQ98qIcwS4Ytmy8A="
+        "CodeSha256": "bOhoKJdCbePXTi0Rpo1weBy21M/B+hxWj/w2NrNVy+0="
       }
     },
     "ApiGatewayRestApi": {
@@ -385,7 +385,7 @@
         "MethodResponses": []
       }
     },
-    "ApiGatewayDeployment1507398194318": {
+    "ApiGatewayDeployment1507410257284": {
       "Type": "AWS::ApiGateway::Deployment",
       "Properties": {
         "RestApiId": {
@@ -474,13 +474,13 @@
     "CreateCompanyLambdaFunctionQualifiedArn": {
       "Description": "Current Lambda function version",
       "Value": {
-        "Ref": "CreateCompanyLambdaVersionDCkSykzItO13HGZDf77sYLdjwclQ98qIcwS4Ytmy8A"
+        "Ref": "CreateCompanyLambdaVersionbOhoKJdCbePXTi0Rpo1weBy21MBhxWjw2NrNVy0"
       }
     },
     "CreateSubscriberLambdaFunctionQualifiedArn": {
       "Description": "Current Lambda function version",
       "Value": {
-        "Ref": "CreateSubscriberLambdaVersionDCkSykzItO13HGZDf77sYLdjwclQ98qIcwS4Ytmy8A"
+        "Ref": "CreateSubscriberLambdaVersionbOhoKJdCbePXTi0Rpo1weBy21MBhxWjw2NrNVy0"
       }
     },
     "ServiceEndpoint": {
