+++ api/event.json
@@ -1,0 +1,7 @@
+{
+   "body" : {
+      "name" : "heh",
+      "email" : "kek",
+      "address" : "awa"
+   }
+}
