--- jobget-service/event.json
@@ -1,7 +1,0 @@
-{
-   "body" : {
-      "name" : "heh",
-      "email" : "kek",
-      "address" : "awa"
-   }
-}
